use @Inject in TvShowsActivity class

+++ app/src/main/java/ir/taghizadeh/dagger2/model/ShowRating.java
@@ -6,5 +6,5 @@ import com.google.gson.annotations.SerializedName;
 public class ShowRating {
     @SerializedName("average")
     @Expose
-    private Integer average;
+    private Double average;
 }
